+++ src/models/knn.py
@@ -21,7 +21,7 @@ class KNEstimator():
         model_params = {k:v for k,v in params.items() if k not in ["cv_folds","beta","scaler","seed","pca","pca_level","scaler_type"]}
         self.classifier = KNeighborsClassifier(**model_params)
 
-    def cross_validate(self, X_train: np.array, y_train: np.array) -> np.float:
+    def cross_validate(self, X_train: np.array, y_train: np.array) -> float:
         splitter = StratifiedKFold(n_splits=self.cv_folds, shuffle=True, random_state=self.seed)
         score = 0.0
         self.cv_scores = {
